training_image/requirements.txt
albumentations
annotated-types==0.7.0
beartype==0.19.0
certifi==2024.12.14
charset-normalizer==3.4.1
colorama==0.4.6
contourpy
cycler==0.12.1
deprecation==2.1.0
eval_type_backport==0.2.2
filelock==3.13.1
fonttools==4.55.6
fsspec==2024.2.0
idna==3.10
imutils==0.5.4
Jinja2==3.1.3
kiwisolver==1.4.7
MarkupSafe==2.1.5
matplotlib
mpmath==1.3.0
numpy
orjson==3.10.15
packaging==24.2
picsellia
picsellia-annotations
picsellia-connexion-services
pillow
pycocotools
pydantic
pydantic_core
pyparsing
python-dateutil
python-dotenv
pytorch-warmup
PyYAML
requests
scipy
semantic-version
simsim
six
stringzilla
sympy
#torch==2.5.1+cu121
#torchaudio==2.5.1+cu121
torchmetrics
#torchvision==0.20.1+cu121
tqdm
typing_extensions
urllib3
xmltodict
zipp
